remove momentum from bars
